Moved layout storage
